requirement.txt
Pillow
djangorestframework
psycopg2-binary
django
djangorestframework-simplejwt
redis
redis-server
celery
django-celery-beat
cors-headers-4.3.1
drf-yasg